refactor(common): establish decimal domain owner
